steel_login

Struct JavaTcpClient

pub struct JavaTcpClient {

    pub id: u64,
    pub gameprofile: AsyncMutex<Option<GameProfile>>,
    pub client_information: AsyncMutex<ClientInformation>,
    pub protocol: Arc<AtomicCell<ConnectionProtocol>>,
    pub address: SocketAddr,
    pub cancel_token: CancellationToken,
    pub outgoing_queue: UnboundedSender<OutboundPacket>,
    pub network_writer: JavaNetworkWriter,
    pub compression: Arc<AtomicCell<Option<CompressionInfo>>>,
    pub server: Arc<Server>,
    pub connection_session: Arc<ServerConnectionSession>,
    pub challenge: AtomicCell<[u8; 4]>,
    pub connection_updates: Sender<ConnectionUpdate>,
    pub connection_updated: Arc<Notify>,
    task_tracker: TaskTracker,

}

Connection for pre-play packets.

Gets dropped by incoming_packet_task if closed or upgraded to play connection.

Fields
id: u64

The unique ID of the client.

gameprofile: AsyncMutex<Option<GameProfile>>

The client’s game profile information.

client_information: AsyncMutex<ClientInformation>

The client’s settings (view distance, language, etc.) received during config.

protocol: Arc<AtomicCell<ConnectionProtocol>>

The current connection state of the client (e.g., Handshaking, Status, Play).

address: SocketAddr

The client’s IP address.

cancel_token: CancellationToken

A token to cancel the client’s operations. Called when the connection is closed.

outgoing_queue: UnboundedSender<OutboundPacket>

A queue of encoded packets to send to the network.

network_writer: JavaNetworkWriter

The packet encoder for outgoing packets.

compression: Arc<AtomicCell<Option<CompressionInfo>>>

Current compression settings.

server: Arc<Server>

The shared server state.

connection_session: Arc<ServerConnectionSession>

The session id state for the active server connection listener

challenge: AtomicCell<[u8; 4]>

The challenge sent to the client during login.

connection_updates: Sender<ConnectionUpdate>

Channel for broadcasting connection state updates.

connection_updated: Arc<Notify>

Notification for when connection updates are processed.

task_tracker: TaskTracker
Implementations

impl JavaTcpClient

pub fn handle_config_custom_payload(&self, packet: SCustomPayload)

Handles a custom payload packet during the configuration state.

pub async fn handle_client_information(&self, packet: SClientInformation)

Handles the client information packet during the configuration state.

pub async fn start_configuration(&self)

Starts the configuration process by sending initial packets.

pub async fn handle_select_known_packs(&self, packet: SSelectKnownPacks)

Handles the select known packs packet during the configuration state.

pub(crate) async fn finish_configuration(&self) -> ConnectionAction

Finishes the configuration process and transitions to the play state.

Panics

This function will panic if the game profile is empty, should be impossible at this point.

impl JavaTcpClient

pub(crate) async fn handle_hello(&self, packet: SHello) -> ConnectionAction

Handles the hello packet during the login state.

Panics

This function will panic if the player name converted to a UUID fails.

pub(crate) async fn handle_key(&self, packet: SKey) -> ConnectionAction

Handles the key packet during the login state, used for encryption.

pub(crate) async fn finish_login( &self, profile: &GameProfile, ) -> ConnectionAction

Finishes the login process and transitions to the configuration state.

Panics

This function will panic if the compression threshold cannot be converted to an i32.

pub async fn handle_login_acknowledged(&self)

Handles the login acknowledged packet and transitions to the configuration state.

impl JavaTcpClient

pub async fn handle_status_request(&self)

Handles a status request from the client.

pub async fn handle_ping_request(&self, packet: SPingRequest)

Handles a ping request from the client.

impl JavaTcpClient

pub fn new( tcp_stream: TcpStream, address: SocketAddr, id: u64, cancel_token: CancellationToken, server: Arc<Server>, connection_session: Arc<ServerConnectionSession>, task_tracker: TaskTracker, ) -> (Self, UnboundedReceiver<OutboundPacket>, TCPNetworkDecoder<BufReader<OwnedReadHalf>>)

Creates a new JavaTcpClient.

pub fn close(&self)

Closes the connection.

pub async fn send_bare_packet_now<P: ClientPacket>(&self, packet: P)

Sends a packet immediately, without queuing.

Panics

This function will panic if the packet cannot be encoded. Should never happen.

pub async fn send_packet_now(&self, packet: &EncodedPacket)

Sends an already encoded packet immediately, without queuing.

async fn write_network_packet( network_writer: &JavaNetworkWriter, packet: &EncodedPacket, ) -> Result<(), PacketError>

async fn release_network_writer(network_writer: &JavaNetworkWriter)

pub fn send_bare_packet<P: ClientPacket>( &self, packet: P, ) -> Result<(), PacketError>

Encodes and queues a packet to be sent.

pub fn send_packet(&self, packet: EncodedPacket) -> Result<(), PacketError>

Queues an already encoded packet to be sent.

pub fn start_outgoing_packet_task( self: &Arc<Self>, sender_recv: UnboundedReceiver<OutboundPacket>, )

Starts a task that will send packets to the client from the outgoing packet queue. This task will run until the client is closed or the cancellation token is cancelled.

async fn write_queued_disconnect( network_writer: &JavaNetworkWriter, sender_recv: &mut UnboundedReceiver<OutboundPacket>, id: u64, )

pub fn start_incoming_packet_task( self: &Arc<Self>, reader: TCPNetworkDecoder<BufReader<OwnedReadHalf>>, )

Starts a task that will receive packets from the client. This task will run until the client is closed or the cancellation token is cancelled.

async fn process_packet( &self, packet: RawPacket, ) -> Result<ConnectionAction, PacketError>

pub async fn handle_handshake( &self, packet: RawPacket, ) -> Result<(), PacketError>

Handles a handshake packet.

pub async fn handle_status(&self, packet: RawPacket) -> Result<(), PacketError>

Handles a status packet.

pub(crate) async fn handle_login( &self, packet: RawPacket, ) -> Result<ConnectionAction, PacketError>

Handles a login packet.

pub(crate) async fn handle_config( &self, packet: RawPacket, ) -> Result<ConnectionAction, PacketError>

Handles a configuration packet.

pub async fn kick(&self, reason: TextComponent)

Kicks the client with a given reason.